Better CRL support
CRLs can now be generated from both Certificates and SerialNumbers.
This is mostly for convenience if you don't want to keep the entire
certificate around or want a lighter-weight way of doing it.

There's also a new Revocable module to organize revocation-centric
attributes.

Additionally, the tests have been shorn up and the testing modulus
reduced to 768 to speed things up.